fix: merge과정에서 발생한 에러 처리
+ properties에 username, password 값이 변경되어 있어 scoula, 1234로 수정했습니다.

+ RootConfig MapperScan에 report.mapper이 빠져있어 추가했습니다.

+ MemberExceptionHandler의 33번줄 handleDuplicateNickname 메소드의 어노테이션이 DuplicateEmailException으로 잘못 설정되어 있어 DuplicateNicknameException으로 수정하였습니다.

+++ backend/src/test/java/org/scoula/settlement/service/SettlementServiceTest.java
@@ -41,7 +41,7 @@ public class SettlementServiceTest {
     @DisplayName("정산 1단계 요약 정보 조회 테스트")
     public void testGetSettlementSummaryForGapyongTrip() {
         // === GIVEN ===
-        Long tripId = 3L;
+        int tripId = 3;
         log.info("🚀 정산 요약 정보 조회 테스트 시작: tripId = {}", tripId);
 
         // === WHEN ===
@@ -78,7 +78,7 @@ public class SettlementServiceTest {
     @DisplayName("최종 정산 결과 계산 테스트 (가평 여름휴가)")
     public void testCalculateFinalSettlement() {
         // === GIVEN ===
-        Long tripId = 4L;
+        int tripId = 4;
         log.info("🚀 최종 정산 결과 계산 테스트 시작: tripId = {}", tripId);
 
         // === WHEN ===
